datapath: Add DSR for NodePort BPF
This commit adds a direct server return (DSR) support for the NodePort
BPF for IPv4 and in the direct routing mode.

The main idea of DSR is to avoid SNAT'ing an original request sent to an
LB, so that a backend could directly reply to a client (the originator
of the request) and the original source IP could be preserved.

To achieve this, we introduce a new IPv4 option which stores a NodePort
service IP and port number. The option is set by bpf_netdev running on
a public iface of an intermediate node which received the original request.
Once the option has been set, the request (the dst IP addr of the request is
DNAT'd to the backend IP addr) is forwarded to a node running the backend.
After receiving the fwd'd request, bpf_lxc of the backend parses the
option, stores the svc addr:port in the NAT table and sets the "dsr" bit
in a CT entry.

When sending a reply to the client, bpf_lxc finds out that the "dsr" bit
was set, does a lookup in the NAT table to find the mapping, and finally
rewrites the source addr and port to the svc addr and port.

The current approach has a shortcoming that if the request size is > (MTU -
8bytes), the request will be dropped after we append the IPv4 option.

To partially solve this, in the case of TCP we set the option only for
SYN packets which should have an empty payload. However, the problem
still exists for TCP with SYN cookies and UDP packets. For those cases,
a client needs to decrease its MTU by 8bytes.

Signed-off-by: Martynas Pumputis <m@lambda.lt>